Create basic application setup with auth flow
Includes things related to launching the app and
fetching user auth details. Fetched data is then to frontend,
so it is ready to be used there.

Auth flow also includes conditions to check token expiration.
If token is expired - user will be asked to authorize app
one more time.

Model for saving user data might be extended in future to
keep track of more user information. For example playlists may be stored
the same way.